db+sqlite+pg: split ResolveOwnedAddresses into its own file
Move ResolveOwnedAddresses into addressstore_resolveowned.go in each backend.
It is a thin batched read (one query plus a map assembly), so per the
wallet/internal/db README it stays a direct backend implementation with no ops
interface.
